[DI-27060] - PR comments
